third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/debugger-scope-resolve-this-expected.txt

CONSOLE MESSAGE: line 7: [object Object]
Tests resolving this object name via source maps.

Set timer for test function.
Script execution paused.

Scope variables sidebar pane:
Local
    this: Foo {}
Closure (Foo.bar)
WindowGlobal
    <section collapsed>
Script execution resumed.
